+++ js/models/main/newsstnadGrid.js
@@ -1,8 +1,8 @@
-import { shuffle } from '../utils/utils.js';
-import { getNewsData } from '../core/apis.js';
-import { paintNewsstand } from '../components/newStandGrid.js';
-import { globalStore } from '../store/globalVarStore.js';
-import { subScribeStore } from '../store/subScribeStore.js';
+import { shuffle } from '../../utils/utils.js';
+import { getNewsData } from '../../core/apis.js';
+import { gridView } from '../../components/gridView.js';
+import { globalStore } from '../../store/globalVarStore.js';
+import { subScribeStore } from '../../store/subScribeStore.js';
 let datas = [];
 
 subScribeStore.subscribe(() => initNewsStandGrid());
@@ -16,7 +16,7 @@ async function initNewsStandGrid() {
 
 function 전체_언론사(newsData) {
   datas = newsDataPaser(newsData);
-  paintNewsstand(datas, globalStore.state.전체언론_그리드_인덱스);
+  gridView(datas, globalStore.state.전체언론_그리드_인덱스);
 }
 
 function 내가_구독한_언론사(newsData) {
@@ -24,7 +24,7 @@ function 내가_구독한_언론사(newsData) {
   datas = newsData.filter((data) => subscribeData.includes(data.name));
   datas = newsDataPaser(datas);
   while (datas.length < 96) datas.push(['', '']);
-  paintNewsstand(datas, globalStore.state.구독언론_그리드_인덱스);
+  gridView(datas, globalStore.state.구독언론_그리드_인덱스);
 }
 
 function newsDataPaser(datas) {
